requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
mutagen>=1.46
python-dotenv>=1.0
gunicorn>=20.1
watchdog>=2.1
